requirements.txt
aiogram
python-dotenv
APScheduler
timezonefinder
psycopg[binary]
